modelmanagement: Implement delete and move routes to cud
