+++ internal/handlers/routes/cud/deleteModel.go
@@ -1,31 +1,16 @@
-package modelmanagement
+package cud
 
 import (
-	"gorm.io/gorm"
-	"io"
 	"net/http"
-	"net/url"
-)
+	"strings"
 
-type DeleteModelRequest struct {
-	ID uint `json:"id" schema:"id,required"`
-}
+	"gorm.io/gorm"
+)
 
 func DeleteModel[T any](w http.ResponseWriter, r *http.Request, db *gorm.DB, model *T) {
 	ctx := r.Context()
 
-	body, err := io.ReadAll(r.Body)
-	if err != nil {
-		http.Error(w, "Failed to read body: "+err.Error(), http.StatusBadRequest)
-		return
-	}
-	values, err := url.ParseQuery(string(body))
-	if err != nil {
-		http.Error(w, "Failed to parse form data: "+err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	id := values.Get("id")
+	id := strings.TrimPrefix(r.Header.Get("AJAX-Target"), "row-")
 	if id == "" {
 		http.Error(w, "ID is required for deleting a model", http.StatusBadRequest)
 		return
